(PUP-4688) Install dmg during AIO setup
Installs dmg on OSX during AIO setup. The `SHA` environment variable
should be the full SHA (or tag) being tested, and should correspond to the
location in http://builds.puppetlabs.lan/puppet-agent/<FULL_SHA_or_tag>.
The `SUITE_VERSION` should be the git describe (or tag) in the filename,
e.g. puppet-agent-<SHORT_SHA_or_tag>-osx-10.10-x86_64.dmg. For example:

    bundle exec rake ci:test:aio \
      SUITE_VERSION=1.1.1.250.gb9e1783 \
      SHA=b9e178350dd52c75ab98ffd05e9bc739d7609155 \
      CONFIG=config/nodes/osx-109-x86_64.yaml

In the case of a tag, e.g. 1.2.1, the SUITE_VERSION can be omitted since
it is always same as the SHA.

This commit also updates the beaker dependency to a version that supports
installing dmg's from something other than latest.