(PUP-4819) Add class info service
This adds a class info service that returns information about classes
and their parameters in files per environment.

To obtain information about classes and their parameters, call
Puppet::InfoService.classes_per_environment(env_file_hash) wher
env_file_hash is a hash with environment name to array of file paths to
.pp files.

The result is a map from environment name, to file, and then
to an array of classes with parameters.

~~
{ "production_env"=>{
  ".../fum.pp" => [
    {:name=>"fum", :params=>[{:name=>"fum_a"}]}
    ]
  },
  "test_env" => { ... }
}
~~~

See ticket for more details.