(PUP-4826) Make Integer and Float with one argument unbounded
The Puppet evaluator will consider this statement to be true:

 Integer[0] == Integer[0,default]

The TypeParser (used in Ruby function API) will instead use the single
argument in both positions and thus consider this to be true:

 Integer[0] == Integer[0,0]

This commit changes the TypeParser to behave like the evaluator. The
change affects both Integer and Float.