(PUP-4490) Pick up bindings from module's metadata.json
This commit adds a 'module:/*::metadata' pattern to the includes
for the binding layer 'modules'. The ModuleScheme will recognize
the matching entries and assume that contributed bindings are
specified in the module's metadata.json file using the key
'data_provider'.

Unit tests are provided to verify that a data_provider that appoints
a Puppet provided binding (such as 'function') works OK and that
a user provided binding (such as 'sample') also works.