(PUP-4810) Make end of time for 32 bit old rubies not crash
Before this, the end of time caused problems on 32 bit machines using
old Ruby versions where time is a 32 bit value. The max time is then the
year 2038. On 64 bit machines, and later Rubies a 64 bit int is always
used and they do not have this problem.

This commit first attempts to get a very very late end of time, and if
that fails backs down to the year 2038.