(RE-5087) Fix init script behavior on sles/suse when run by puppet itself
On suse/sles 10, when executed by puppet, the init script will not start the puppet
service because startproc checks and finds a puppet process running.

This updates the init script to force start the process unless there is
a running process matching the pid in the pidfile. It also updates the status
command to only return true if the pidfile matches a running puppet pid.

This is a direct port of a fix for SLES 10 from previous PE releases.