(PUP-4919) static compiler always sets ensure parameter to 'file' if a source is declared
The static_compiler catalog terminus always sets ensure to 'file' on
file resources as long as the resource has a source parameter declared.
It does this even if the ensure parameter is explicitly declared.

This is inconsistent with the behavior of the normal compiler, where a
file's ensure parameter gets enforced regardless of the value of the
source parameter.

This commit skips file resources with ensure set to 'absent' in the static
compiler.