(PUP-4926) Make it possible to form relationship with 3x Resource
Before this it was not possible to form a relationship with a 3x
Resource instance. This problem occurs if passing a PResourceType
to a defined type or class. The evaluation transforms the PResourceType
to a Puppet::Pops::Parser::Resource which is given as argument value.

When later, the defined resource (or class) is evaluated, a relationship
could not be formed using this variable.

The problem was that the Relationship Operator implementation did not
transform the Parser::Resource back to a PResourceType - instead it
failed with an error that Resource was not a CatalogEntry.