(maint) Fix file bucket dipper spec on OSX.
The list command specs were using the path returned from `tmpfile` in
its expectations, but the filebucket returns the real paths to the file.
For OSX, `tmpfile` is returning paths under `/var/...` which is
symbolically linked to `/private/var/...`.  Thus the expectations would
not match.

The fix is to simply get the real path to the files returned from
`tmpfile` and friends.