(PUP-4941) Add warnings for app management keywords
This adds warnings when application managemnt keywords 'application',
'produces' and 'consumes' are used as bare words.

In order to support this, it was not enough to add them as Reserved, as
that means that an error is raised. Instead, the ReservedWord AST model
object got a flag, to describe that a token is reserved "in the future".
The checking logic now checks this flag, and emitts a warning or an
error.

Further, in preparation for more work, the new tokns were added as
APPLICATION_R, PRODUCES_R, etc. to reflect that they are reserved. In
the grammar, these tokens leads to the construction of a ReservedWord
AST element. Later, when a feature switch is introduced, non _R variants
of the tokens are delivered instead of the _R. This means that the
grammar can use APPLICATION token where it has meaning, and
APPLICATION_R for the reserved case (i.e. the grammar is capable of
handling both cases).