(PUP-4485) Replace Scope parameter with a LookupInvocation parameter
This commit introduces a new class named LookupInvocation. This
invocation contains the scope, override values, and default values.
The instance is also responsible of trapping an endless recursion
via its 'check' method.

The intended use is that an instance is created at top level, i.e.
where the lookup originates (typically in the lookup function or
when the Resource is resolving values). The instance is then just
passed on everywhere and never recreated. The instance is passed
instead of the former three parameters (scope, override_values, and
default_values). Each new lookup that happens (currently none, but
we will have recursive lookups when we introduce hiera interpolation).

The current design replaces the old 'scope' parameter with the new
'lookup_invocation' parameter and should have no effect on existing
providers unless they access the scope since the number of parameters
remains the same.