(maint) Fix file bucket spec timing issue.
The file bucket list command feature added a spec that adds three files
to the bucket and compares what is listed with Time.now.  Unfortunately,
this introduces a timing issue where TIme.now is not the same second as
the modification times of the files that were stored in the file bucket.
This results in a sporadic failure.

The fix is to simply verify that the list contents is in the expected
format with respect to the date and time.