(PUP-4905) Add app_managment boolean setting
This adds the boolean setting `app_managemnt` that is used as a feature
switch for application management fatures.

Since it is unclear if this is for both agent and master, it is set in
the main section (in a way that is easy to change to another section if
main should be the wrong choice).

As this only adds the flag, there is not much that can be said in the
description. Later works should add to the description of the setting.