(PUP-4905) Add feature switch for application management tokens
After this, it is possible to turn on the application management tokens
with the `app_management` feature flag. If the flag is off, they are
delivered in the reserved bare word form and will produce deprecation
warnings (i.e. APPLICATION_R, CONSUMES_R, and PRODUCES_R are produced as
tokens). When the switch is on, the same tokens are APPLICATION,
CONSUMES, and PRODUCES.

The above enables a grammar to define both ruls for reserved future
keywords and live rules at the same time.