(PUP-1189) declare report handlers with overwrite
Previously, if a report handler failed during initialization, it wasn't
cleaned up. So the second time a report was submitted, Puppet would try
to instantiate a new class, and then fail because it had already been
instantiated. This instructs the class loader to overwrite if needed.

(PUP-1189) add spec test

(PUP-1189) Corrected spec test for accuracy

Thanks for the help, @MikaelSmith