(PUP-4622) Fix acceptance test in Jenkins CI
In the Jenkins environment, the master options being passed cause the
test to fail when attempting to kill the master. Use default options to
setup the master for the test.

The test would still setup a master, even if no agents were identified
to test (i.e. Windows). Skip the test if no agents are available after
the confine.

Also cleanup some magic and formatting.

[skip ci]