(PUP-3045) Send a TERM when Exec processes timeout
This patch sends a TERM signal to POSIX subprocesses spawned by Exec resources
when the `timeout` paramenter triggers an exception. Any process that does not
respond to TERM before Puppet exits will be orphaned and left for init to clean
up --- this is the default behavior for all timeout events prior to this patch.