(PUP-4824) Fixup acceptance group / user tests for OSX
Previously, several user and group acceptance tests dependeded
on `getent` to create and query users and groups on the system.
This did not work in OSX, and caused several test failures.

We now have cross-platform beaker methods specifically meant for
user and group management which we can use to replace the bare
`getent` calls.