(PUP-4485) Add a 'hiera' data provider
This commit adds 'hiera' data providers for environments and modules.

The 'hiera' data provider will delegate all lookups to other data
providers declared in its 'hiera.yaml' configuration file. Such
data providers are assumed to be derived form the PathBaseDataProvider
class and created using a PathBaseProviderFactory class that is bound
to the name declared as 'backend' in the hiera configuration. At
present, two such factories are bound. One for 'yaml' and one for
'json'.