(maint) Fix bug in stat function
This is coming from a CI failure for puppet-server, which uses puppet
4.2.1, see SERVER-832 for details.

The stat function references an agent variable, which will coincidentally
be set correctly if the beaker config only has one host with an agent role.
But if there is more than one agent, the 'agent' variable will be a list
of hosts, and the stat function will throw an exception.

The bug was introduced in 8d2a872