(PUP-5009) Use introspected modulepath on failing PMT tests
This commit adds a helper method to determine a non-default
moduledir in the available module path. This method is then used in
conjunction with the get_default_modulepath_for_host method to
replace 'sitemoduledir' and 'distmoduledir' in tests that were
failing due to these values not being on the modulepath.