(PUP-5014) Base class for JSON/msgpack indirection termini calls log_exception() wrong
Puppet::Indirector::JSON and Puppet::Indirector::Msgpack call Puppet.log_exception() without
a mandatory exception as the first argument value, passing the message string instead.

This breaks further along in format_exception() when it attempts to get
the exception message attribute but finds a string, resulting in a
NoMethodError raised by the puppet agent.

This commit fixes the issue by passing the raised exception to
log_exception() as required first, and the message second.