(PUP-5035) undefined method `keys' for nil:NilClass in static_compiler
get_child_resources() can return nil if the source attribute of the file
resource points to a file instead of a directory.

This results in a raised NoMethodError with the static compiler set as
the catalog terminus and a file resource declared that has the recurse
parameter set to true.

This commit prevents the attempt to add child resources to the catalog when
there are none.