(maint) Only install msi and dmg once
Prior to Beaker 2.19, puppet's acceptance tests installed puppet-agent
twice on Windows and OSX. On Windows, the first install did nothing, and
only when we installed the second time did it work:

    $ cmd.exe /c start /w C:/PROGRA~3/puppet-agent-x86.msi
    ...
    $ cmd.exe /c start /w msiexec /qn /i \
      http://builds.puppetlabs.lan/puppet-agent/be1f554d82b2f6a8ee335e97d102e62209155266/artifacts/windows/puppet-agent-x86.msi \
      /L*V C:\\Windows\\Temp\\Puppet-Install.log

On OSX, it installed successfully the first time and did an upgrade the
second time:

    $ installer -pkg /Volumes/puppet-agent-1.2.1.119.gbe1f554*/puppet-agent-1.2.1.119.gbe1f554*.pkg -target /
    installer: Package name is puppet-agent
    installer: Installing at base path /
    installer: The install was successful.
    ...
    $ installer -pkg /Volumes/puppet-agent-1.2.1.119.gbe1f554*/puppet-agent-1.2.1.119.gbe1f554*.pkg -target /
    installer: Package name is puppet-agent
    installer: Upgrading at base path /
    installer: The upgrade was successful.

It turns out that Beaker's `install_puppet_agent_dev_repo_on` method
installs puppet-agent on platforms that don't have a package manager, as
opposed to doing a no-op.

Beaker 2.19.0 was modified to install the MSI as a manual service by
default, and the `install_puppet_agent_dev_repo_on` method was modified
to correctly install the MSI. When the `010_Install` step installed the
MSI the second time, it preserved the service state, and tried to stop
the service, but failed because the service was already stopped.

This commit bumps our Beaker dependency to 2.19.0, which correctly
installs the puppet-agent MSI and DMG. It eliminates the second install
on Windows and OSX. It also moves the logic for verifying the puppet and
ruby versions so it is executed on all hosts, not just windows.

[skip ci]