(PUP-5041) "environment non existent" acceptance tests failing on puppet-server master
This change ensures we run the apply, module, config tests on the master
even if the master does not have an agent role.  We removed the agent
role from the master in the puppet tests to decrease testing time at low
risk.
This change also fixes the environment_scenario-bad test which never
would have passed against current puppet nor server due to matching on
output that was fixed.

[skip ci]