(PUP-5002) acceptance: add metadata.json data binding to lookup.rb test
The existing test creates manifests including the bindings directory in the module's lib/puppet/ directory. The additional test code creates new manifests which remove the bindings directories and install a metadata.json file for each of the two modules. The new manifest is applied and the asserts are checked
again to verify the same data is returned as was before when the data bindings were made in the module's lib/puppet/bindings.