(PUP-5051) Add --no-check-prereqs flag to install.rb
In cases where the ruby used to run install.rb is not the ruby that will
be running puppet, or where the system/platform that install.rb is run
on will not be the system/platform that puppet will be running on,
checking for and using facter in the install is less than useful. This
commit adds a flag to support not checking for prerequisite libraries
during install. This commit also moves the operatingsystem check for
building RI into the method itself, to help facilitate the reordering
here.