(PUP-1455) Use CFPropertyList to read launchd plists
This uses the CFPropertyList gem which is shipped with Puppet to read launchd
plists instead of shelling out to plutil to convert them and then reading
them in. This saves a lot of time, between 10 and 20 seconds per Puppet run.

This commit introduces a Puppet::Util::Plist namespace for handling plists
with CFPropertyList and modifies the existing launchd service provider to
use it. Activities in Puppet::Util::Plist that will use common libraries
have been wrapped in their own helper methods (which will allow us to mock
out those methods for testing purposes). This level of encapsulation will
allow us to swap out the underlying implementation should things change in
the future.