(PUP-3318) Change environment change warning to notice
When the puppet agent is instructed to change environments by
classification, it issues a warning. That change is expected, so a
warning is not.

Change the warning to a notice. Later environment changes caused by
reclassification based on new facts are still warnings, as changing
environments multiple times isn't generally desirable.