(PUP-5074) Support consume/produces annotations site-wide
This commit ensures that the resolution of declared capability
mappings is deferred until all files for a particular environment
has been parsed.

The commit introduces a new Resource::Type named 'capability_mapping'
that acts as intermediate object created by the parser and added to a
hash in the known_resource_types TypeCollection which is later made
available to the compiler. The compiler uses these objects when it's
time to evaluate to perform the actual mapping between capabilities
and resources. Once the mapping is complete, the 'capability_mapping'
hash is no longer needed and is cleared out to conserve memory.