(PUP-5032) Add info message identifying the environment used
Previously Puppet's log output would not identify the environment used
unless it was changed by a node classifier. This did not instill
confidence in which environment was used, requiring users to look at
puppet.conf or the run report.

Adds an info message identifying which environment is used for
pluginsync/compilation if the configured environment is used.