(PUP-4485) Better handling of syntax errors and typos fixed.
This commit fixes a couple of typos. It also adds a rescue clause
around the actual parsing of the yaml and json data files to ensure
that a proper Puppet::DataBinding::LookupError is thrown in case
of failures.

Two tests were added to ensure that the parse errors are generated
correctly.