(PUP-5088) Ensure only => can be used in capability mappings
This commit relaxes that constraint that a CapabilityMapping
only can contain an AttributeOperation so that it instead can
contain an AbstractAttributeOperation. This was done to prevent
a cryptic error message when an AttributesOperation (note the
plural) was assigned due to use of *=> operator. The relaxed
constraint lets this pass so that it instead can be trapped
in the validator where a correct check was already in place.