(PUP-4697) Make systemd the default service provider for all suse's *except* 10 and 11
Previously, the default service provider for suse was 'redhat' with
an override of 'systemd' for suse 12 and 13. However this doesn't work for the rolling
release of suse (tumbleweed) or the next (leap), which are both systemd
based.

This reverses the logic so the default service provider is 'systemd'
with overrides for suse 10 and suse 11.

Thanks to Robert Monteanu for researching what suse versions are using what
init systems.

I also updated some unrelated spec tests in provider/service/systemd_spec.rb
to use 'stubs' rather than 'expects' - the latter made the tests unnecessarily
brittle and was testing implementation rather than behavior.