(PUP-5197) Defer appmgmt-specific setup of lexer tables
The app_management feature flag governs which of two variants of lexer
tables we use for appmgmt specific language constructs. Previously, the
decision was made when code was loaded, which is too early for how
puppet-server loads the Ruby code.

We now defer the decision until we actually get ready to lex text, and make
the decision in the initvars method, rather than in the setup of constants
during loading of the Ruby code.

As an (untested) side-effect, it should now be possible to turn
app_management on or off per environment rather than just globally.

This change also required changing before(:all) to before(:each) as
Puppet[:app_management] is reset before each test automatically. On the
upside, we can now also get rid of the horrible hacks in
ParserRspecHelper::with_app_management