(PUP-5078) Validate export/consume metaparameters
This commit adds validation of values assigned to the export/consume
metaparameters. The validation ensures that the value is a capability
resource. It also asserts that the value is a resource before attempts
are made to use methods such as resource.type. This to avoid cryptical
"undefined method" errors in case user enters non resource RHS values.