(PUP-5079) Check that 'nodes' parameter for applications is sound
This commit adds validation of the 'nodes' parameter in application
instances such that it must be a Hash where each key is a Node and
each value is a resource or an array of resources.

The code also checks that no resource is mapped to more than one
node.

A change was made to remove the prior more relaxed scheme where the
node could be on either side of the association in the hash. Now
it must be the key. This is natural since the node must be unique
and since the resources that it is associated with can be an array.