(PUP-4476) Call lookup with the correct parameters
Prior to this commit, the lookup application was calling the internal
lookup function in a way that was compatible with the old version
of the function (eight parameters) rather than the current version
of the function (six parameters). This meant that attempts to use
the lookup application would fail due to the function being called
improperly.

Fix this issue and also remove a test that had too much stubbing
and was not longer useful.