(PUP-4442) Only rmdir if mkdir succeeded
Previously, if you tried to manage a file with content and the parent
directory didn't exist, puppet would report `rmdir` failed:

    Error: Could not set 'file' on ensure: No such file or directory @
    dir_s_rmdir - .../file20150909-11354-3uzm9m.lock

The file type uses `Puppet::Util.replace_file` to atomically overwrite
file content, which creates a unique file in the same directory as
the file we're trying to manage, to ensure we're on the same filesystem.
But if the parent directory doesn't exist, `mkdir` raises ENOENT, and
then we called `rmdir`, which also raises ENOENT, masking the original
exception.

This commit changes uniquefile to only call `rmdir` if `mkdir` succeeds.