(PUP-2575) Sort OSX group membership before comparison
OSX group membership is currently returned as simple Arrays. Comparing these
is strict, and the order of the resource must exactly match the order given
in the manifest, or a spurious change notice will be generated.

To fix this, add a members_insync? method on the directoryservice group
provider. This method sorts and uniqs the member lists.

We ensure current and should both respond to :sort. This avoids an error when
an existing group with no members is changed to have members.

If members is [], members_insync? is called with current as :absent. This may
be a problem with the provider, but for now let's handle this case in
members_insync? and return true if current is :absent and should is empty.