(PUP-4475) Add explain-ability to the lookup and data provider APIs
This commit adds report callbacks to the Lookup::Invocation object
which is passed down to all instances involved in the lookup. These
callbacks will either do nothing at all (no Explainer attached to
the Invocation) or tell the explainer to accept data for that
particular report.

A new 'with' method was also added. It is used when a lookup changes
its context. This happens when the lookup is branching due to things
like several paths, global/environment/module scopes, or when
evaluating interpolations in a string. The with method is used as
a wrapper (yields to a block where the lookup continues) and will
push and pop contextual objects in the Explainer and thereby build
a full DAG of all branches that the lookup invocation took part in.

The LookupInvocation that resided in Puppet::DataBinding was moved
to Puppet::Pops::Lookup and renamed to Invocation to make it adjacent
to the Explainer. This was motivated by their close affinity and that
neither of those classes are specific to DataBinding. The impact of
this will be very minor given that no release has gone out where the
LookupInvocation has been included.