(PUP-4475) Improve ability to control output rendering
This commit adds a 'render-as' command line option. The format
will be mapped to one of the renderers registered with
`Puppet::Network::FormatHandler` which then will be used when
rendering the result.

The default format will be 'yaml' when rendering the result of a
lookup and 's' (text) when rendering an explanation.