(PUP-4475) Change option names to use dash instead of underscore
This commit unifies all options to use dashes instead of underscores
in the option name since that seems to be more consistent with how
other options are declared in Puppet.

The commit also contains some minor refactorings such as using the
`main` method instead of `run_as` (declared to call main in the super
class) and name change of RUNHELP to RUN_HELP. Also 'keys' is now a
 local variable since there's no need to put it into the 'options'
hash.