(PUP-4695) New acceptance test for hiera function calls from with templates
Until now we had no acceptance tests for verifying behavior of hiera
function calls (hiera(), hiera_array(), hiera_hash(), and hiera_include()) f
within a template (erb or epp).

This test calls hiera functions from with in an erb template and from within
an epp template.  The master is configured with a manifest.  The client
applies the manifest using the puppet agent -t call.  The file created by
templates are examined to determine whether the hiera function calls returne
the expected values.

hiera_include() is called from the epp template.  The included sub-modules
each create a file that is then checked for correctness.