(PUP-5258) Unquote PUPPET_EXTRA_OPTS in SUSE init script
Prior to this commit, the Puppet service fails to start
on SLES11 with the error "Could not prepare for execution:
The puppet agent command does not take parameters".

This appears to be caused by an empty '' at the end of
Puppet's SUSE init script startproc call.

Since PUPPET_EXTRA_OPS is set to an empty string in the
default /etc/sysconfig/puppet, we update the SUSE
init script to remove the double quotes from
"${PUPPET_EXTRA_OPTS}".