(PUP-5228) Fixup ssh_authorized_key tests
Previously, the ssh_authorized_key tests would append a new
key to authorized_keys with `echo.` On CentOS 4. simply appending
did not create a newline, and so the key was appended to the end
of the last key, meaning Puppet couldn't find or remove it.

This commit updates the test to add a newline to the file before
appending the new key.

In addition, there was an issue with the modification test in which the
key was appended improperly, and we then mistook Puppet's newly generated
key for a modified version of the old key, when in reality
'ensure=present' was just adding a brand new key.