(PUP-5257) Update client init script for EL4
Prior to this commit, managing the AIO
puppet-agent's puppet service using
`puppet resource` fails on EL4.

Because EL4's /etc/init.d/functions do not
take pid as an argument, we're not able to
use the built-in 'status' or 'killproc'
functions, as these will check the status of
and/or kill the `puppet resource` process
itself.

This commit updates the EL init script to
allow use of `puppet resource` to manage the
puppet service on EL4, while continuing to
use the built-in 'status' and 'killproc'
functions on EL5 and EL6.