(maint) Don't use confine_block to skip tests
Previously, we were using beaker's confine_block method to skip symlink
tests on 2003. Prior to beaker 2.24, the test would be skipped on all
platforms due to BKR-505. Beaker 2.24 fixed the issue, which causes the
test to execute correctly on non-2003 platforms, but fails on 2003. That
issue is filed as BKR-535.

This commit changes the acceptance tests to not use `confine_block` to
avoid any beaker incompatibilities.

[skip ci]