(PUP-5192) Add status event to generated report
Previously the round-trip through yaml test passed but didn't exercise
the code that caused PUP-5192, because the generated report didn't
contain any events. This commit updates the test to add an event to the
report. Since events are not value objects, round-tripped events will
never be equal to the expected events, so we compare data hashes
instead.