(PUP-5087) Add evaluation of site expression
This adds evaluation of the site expression. It is treated analogous to
a node expression - but it contains the resources for all cross
node resources as apposed to the resources for a specific node.

This changes the environment compiler to prune the resulting catalog
from unwanted resources instead of attemting to stop them from being
added.

This implementation allows both the pre-site-expression style (in top
scope) and that application/cross node resources are in a
site-expression. This is expected to change.