Revert "(maint) Optimize adding tags"
This reverts commit 0412515e2f9d75f00687bc5c01915da5a51c58e3.

This change to tag filtering meant that a single tag like foo::bar
would be split on the :: seperator and expaned into:
['foo::bar', 'foo', 'bar']. This meant that an agent would sync more
resources than the used intended when tag filtering is in effect.

In order to fix this problem, revert the commit that initially made
this change to filtering behavior.