(PUP-4695) Fix for hiera_in_templates acceptance test
As originally checked in, hiera_in_templates.rb caused failures in CI for
agents whose OSs were Windows or Ubuntu.  The OS family fact, used for
creating hiera data filese, was resolved for the master only.  If the
agent had an OS different from the master's (e.g. Ubuntu or Windows), it
would not find its corresponding hiera file.  The test now polls the agents
and creates a hiera data file for each agent OS family.

In addition, the owner and group attributes were removed, because 'root'
is not recongnized in Windows.

Original check in comment:

Until now we had no acceptance tests for verifying behavior of hiera
function calls (hiera(), hiera_array(), hiera_hash(), and hiera_include())
from within a template (erb or epp).

This test calls hiera functions from with in an erb template and from within
an epp template.  The master is configured with a manifest.  The client
applies the manifest using the puppet agent -t call.  The file created by
templates are examined to determine whether the hiera function calls returned
the expected values.

hiera_include() is called from the epp template.  The included sub-modules
each create a file that is then checked for correctness.