(PUP-5274) Create settings related files with the correct mode
Previously, puppet would create settings related files using the process
umask, instead of the mode specified in defaults.rb, or overridden in
puppet.conf, and it would take a second puppet run to converge.

This issue was introduced in 691056e for PUP-2349, merged in 87b0b5d,
and has been present since 3.7.0. There are two issues with the code:

First, the case statement was trying to compare if mode was a string,
nil, or fixnum using ruby's case equality operator, but it was
actually comparing if mode's class was a string or nil, which was
never true, always falling through to the else case.

Second, the assignment statement `mode = case mode` introduced a `mode`
local variable whose value was set to nil, and shadowed the `mode`
accessor. This combined with above issue meant the method always yielded
nil, and Puppet would eventually call `File.open` with a nil mode when
creating the settings related file.

There is a DirectorySetting class that extends FileSetting, but it
doesn't have this problem because it uses puppet's file type to create
the directory, correctly passing mode to the `Dir.mkdir` call.

This commit avoids the shadow problem, and tests if `mode` is a String.
It also adds unit and acceptance tests as none of the existing tests
failed when the code was changed.