(PUP-5272) Allow test to pass if ctime/mtime don't converge
Previously, the test required all checksum types to converge in one run.
However, ctime and mtime are sensitive to clock skew, and can require
multiple runs to converge. For example, suppose the source file's mtime
on the master is 11:00, but the agent's clock is behind at 10:00. The
agent downloads and applies a catalog, creating the file at 10:02. The
next time the agent runs it will compare its local time of 10:02 against
11:00, decide the file is out of sync, and update it again. This process
will repeat until the file's ctime on the agent is greater than or equal
to the file's ctime on the master.

This commit modifies the install step to query each host for its local
time, and does a rough calculation to determine the difference between
all hosts. Note the timestamp that beaker outputs is on the coordinator
node and so is useless when trying to correlate timestamps.

This commit also modifies the file source attributes test to allow exit
codes 0 and 2 during the second run, ensure md5* and sha* checksums do
converge, and warn, but not fail, if ctime and/or mtime don't converge.

[skip ci]