(PUP-4055) Add DNF provider
This commit adds a new DNF package provider, which is necessary now
that Fedora has moved away from Yum. Since dnf is nearly identical
to yum in functionality, this provider simply inherits all of yum's
methods, changing only the name of the binary and the error level
option to be used when installing packages.