(PUP-5299) Flatten the code structure produced by the 4x code merger
Before this, if a site used thousands of manifests the code merger would
create a tree that was thousands of levels deep. When evaluating that,
the ruby stack will (most likely) hit its limit.

After this change the structure is flattened - rather than recursing
the merged instructions are simply concatenated (after unwrapping them
as much as possible).

As a consequnce, one test was changed since it depended on the internal
structure being build having a particular shape.