(PUP-5016) Disable Init provider on Debian and Ubuntu
Previously, Debian 8 and Ubuntu 15.05, which are systemd-managed
platforms, would erroneously fall back to the init provider
when Systemd was unable to manage a SysVInit service without
a unit file. The Debian provider should be used instead, as
it inherits from init and provides the ability to enable and
disable services.