(PUP-5016) Update linux services tests for Debian systemd behavior
Previously, services_enable_linux.rb was failing in Debian 8
and Ubuntu 15.05 due to issues with Systemd and SysVInit. The
Debian 8 issues are corrected by the actual fix to the systemd
and init providers, but the test continued to fail in Ubuntu
due to an upstream systemd-sysv compatibility bug.

This commit updates the test to use a service which has
a systemd unit file, cron, rather than apache2 which did not. This
allows the test to run correctly on Ubuntu. Once this upstream bug
is fixed, we'll probably want to undo the Ubuntu changes for
consistency.

In addition, the systemd masking test is also updated for strange
or broken Debian and Ubuntu systemd / Init behavior. In this case,
Ubuntu suffers from the same bug as above. In Debian, Apache2 does
not create the expected systemd symlinks as normal systemd services
do, so we use cron.service instead.

See https://bugs.launchpad.net/ubuntu/+source/systemd/+bug/1447807
for the upstream bug report.