(PUP-5016) Fix issue with Debian provider and SysVInit-Systemd compatibility layer
With the compatibility layer between systemd and SysVInit in Debian,
services can be managed in multiple ways. However, if a service is
masked with systemd, the equivalent SysVInit script becomes unable
to determine the status of the service. If we are trying to manage
the SysVInit version, we need to use systemctl to determine the
status in Debian 8 and Ubuntu 15.04.

Note that this is an issue only if querying the SysVInit version
of a service. `puppet resource service ntp` will fail to get the
status, while `puppet resource service ntp.service` will not.